# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for a Todo App",
        "title": "Todo API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:9000",
    "basePath": "/todo",
    "paths": {
        "/item": {
            "get": {
                "description": "Get todo items from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo-item"
                ],
                "summary": "Get todo items",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/todoitem.TodoItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "description": "Create todo item in the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "todo-item"
                ],
                "summary": "Create todo item",
                "parameters": [
                    {
                        "description": "body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/todoitem.TodoItem"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
        "todoitem.TodoItem": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                }
            }
        }
    }
}